requirements.txt
# Core ML/DL
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Azure ML
azureml-core>=1.53.0
azureml-pipeline>=1.53.0
azureml-pipeline-steps>=1.53.0
azureml-dataset-runtime>=1.53.0
azure-storage-blob>=12.19.0

# Federated Learning
flower>=1.5.0
syft>=0.8.0

# Blockchain
web3>=6.11.0
eth-hash>=0.5.2
eth-utils>=2.3.0

# API
flask>=3.0.0
flask-cors>=4.0.0
flask-restful>=0.3.10
pydantic>=2.4.0

# Database
pymongo>=4.5.0
redis>=5.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.66.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0

# Logging & Monitoring
tensorboard>=2.14.0
wandb>=0.15.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
